requirements.txt
# Novel Translation Pipeline Dependencies

# Core dependencies
ollama>=0.1.0
pyyaml>=6.0

# Testing
pytest>=7.0.0

# Optional but recommended
chardet>=5.0.0
